refactor(todo-api): split monolithic main.rs into modules
