+++ examples/todo-api/src/handlers.rs
@@ -1,0 +1,51 @@
+use modo::HttpError;
+use modo::JsonResult;
+use modo_db::Db;
+use serde_json::{Value, json};
+
+use crate::entity::todo;
+use crate::types::{CreateTodo, TodoResponse};
+
+#[modo::handler(GET, "/todos")]
+async fn list_todos(Db(db): Db) -> JsonResult<Vec<TodoResponse>> {
+    use modo_db::sea_orm::EntityTrait;
+    let todos = todo::Entity::find()
+        .all(&*db)
+        .await
+        .map_err(|e| modo::Error::internal(format!("Failed to list todos: {e}")))?;
+    Ok(modo::Json(
+        todos.into_iter().map(TodoResponse::from).collect(),
+    ))
+}
+
+#[modo::handler(POST, "/todos")]
+async fn create_todo(
+    Db(db): Db,
+    input: modo::validate::Json<CreateTodo>,
+) -> JsonResult<TodoResponse> {
+    input.validate()?;
+    use modo_db::sea_orm::{ActiveModelTrait, Set};
+    let model = todo::ActiveModel {
+        title: Set(input.title.clone()),
+        ..Default::default()
+    };
+    let result = model
+        .insert(&*db)
+        .await
+        .map_err(|e| modo::Error::internal(format!("Failed to create todo: {e}")))?;
+    Ok(modo::Json(TodoResponse::from(result)))
+}
+
+#[modo::handler(DELETE, "/todos/{id}")]
+async fn delete_todo(Db(db): Db, id: String) -> JsonResult<Value> {
+    use modo_db::sea_orm::{EntityTrait, ModelTrait};
+    let todo = todo::Entity::find_by_id(&id)
+        .one(&*db)
+        .await
+        .map_err(|e| modo::Error::internal(format!("Failed to find todo: {e}")))?
+        .ok_or(HttpError::NotFound)?;
+    todo.delete(&*db)
+        .await
+        .map_err(|e| modo::Error::internal(format!("Failed to delete todo: {e}")))?;
+    Ok(modo::Json(json!({"deleted": id})))
+}
